+++ example/src/pages/Components/Globals/TEPopupContextSection.js
@@ -9,7 +9,7 @@ const TEPopupContextSection = () => {
 	return (
 		<ComponentSection>
 			<ComponentTitle
-				title='TEPopupContext & TEPopupProvider'
+				title='TEPopupContext, TEPopupProvider & useTEPopups'
 				link='https://github.com/SCasarotto/react-tec/blob/master/src/Globals/TEPopupContext/TEPopupContext.js'
 			/>
 			<p>
@@ -30,6 +30,21 @@ const TEPopupContextSection = () => {
 			<iframe
 				src='https://codesandbox.io/embed/fzd72?fontsize=12&hidenavigation=1&module=%2Fsrc%2FExample.js'
 				title='TEPopupContext & TEPopupProvider'
+				style={{
+					width: '100%',
+					height: 500,
+					border: 'none',
+				}}
+				sandbox='allow-modals allow-forms allow-popups allow-scripts allow-same-origin'
+			/>
+			<p>
+				In addition, <Code>useTEPopups</Code> is a nice helper hook that provides easy to
+				use funtions instead of the dispatch types. The below example is identical to the
+				one above but uses the <Code>useTEPopups</Code> hook.
+			</p>
+			<iframe
+				src='https://codesandbox.io/embed/fol5p?fontsize=12&hidenavigation=1&module=%2Fsrc%2FExample.js'
+				title='useTEPopups'
 				style={{
 					width: '100%',
 					height: 500,