+++ src/Layout/TEBodyContainer/styledComponents.js
@@ -1,13 +1,9 @@
 import styled from 'styled-components'
 
 export const BodyContainer = styled.div`
-	${(props) => {
-		const { sidebarWidth } = props
-		return `
-            position: relative;
-            width: calc(100% - ${sidebarWidth}px);
-            min-height: 100%;
-            margin-left: ${sidebarWidth}px;
-        `
-	}}
+	position: relative;
+	width: calc(100% - ${(props) => props.sidebarWidth}px);
+	min-height: 100vh;
+	margin-left: ${(props) => props.sidebarWidth}px;
+	background-color: ${(props) => props.theme.lighterGray};
 `